+++ src/Modules/Workflows/Models/StepTypesModel.php
@@ -14,6 +14,7 @@ use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\Deact
 use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\DeletePost;
 use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\DeletePostMeta;
 use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\DoAction;
+use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\InteractiveDelay;
 use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\QueryPosts;
 use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\RemovePostTerm;
 use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\ScheduleDelay;
@@ -262,6 +263,7 @@ class StepTypesModel implements StepTypesModelInterface
             Conditional::getNodeTypeName() => new Conditional(),
             AppendDebugLog::getNodeTypeName() => new AppendDebugLog(),
             DoAction::getNodeTypeName() => new DoAction(),
+            InteractiveDelay::getNodeTypeName() => new InteractiveDelay(),
         ];
 
         if (function_exists('ray')) {