+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add new workflow trigger "On custom action" that allows workflows to be triggered by custom WordPress action hooks, enabling integration with other plugins and custom code - PRO feature (Issue #1222).
 - Add new workflow step "Do action" that executes custom action hooks with arguments, enabling integration with other plugins and custom code - PRO feature (Issue #1222).
-- Added diagnostic check for Spatie Ray debugging tool in the Diagnostics and Tools settings tab, clarifying debugging capabilities.
-- Added a "Copy" button to workflows (Issue #1183).
-- Added a "Cancel Scheduled Actions" button to workflows lists (Issue #1326).
+- Add diagnostic check for Spatie Ray debugging tool in the Diagnostics and Tools settings tab, clarifying debugging capabilities.
+- Add a "Copy" button to workflows (Issue #1183).
+- Add a "Cancel Scheduled Actions" button to workflows lists (Issue #1326).
+- Add Add a new step for interactive delay (Issue #1257).
 
 ### Changed